ml/requirements.txt
openai-whisper
torch
torchvision
torchaudio
numpy
matplotlib
scikit-learn
nltk
ffmpeg-python
pandas